Add GlobalPrefixHints for SDL bindings and regenerate on Linux
VaListTagHandle is an extra handle struct that seems to generate on Linux (and not on Windows) and breaks global prefix determination.

+++ sources/SDL/SDL/SDL3/DropEvent.gen.cs
@@ -16,7 +16,7 @@ public unsafe partial struct DropEvent
     public uint Reserved;
 
     [NativeTypeName("Uint64")]
-    public ulong Timestamp;
+    public nuint Timestamp;
 
     [NativeTypeName("SDL_WindowID")]
     public uint WindowID;
