+++ sources/SDL/SDL/SDL3/InitState.gen.cs
@@ -3,7 +3,6 @@
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace Silk.NET.SDL;
 
@@ -12,6 +11,6 @@ public unsafe partial struct InitState
     public AtomicInt Status;
 
     [NativeTypeName("SDL_ThreadID")]
-    public ulong Thread;
+    public nuint Thread;
     public void* Reserved;
 }
