+++ sources/SDL/SDL/SDL3/PenTouchEvent.gen.cs
@@ -16,7 +16,7 @@ public partial struct PenTouchEvent
     public uint Reserved;
 
     [NativeTypeName("Uint64")]
-    public ulong Timestamp;
+    public nuint Timestamp;
 
     [NativeTypeName("SDL_WindowID")]
     public uint WindowID;
