+++ sources/SDL/SDL/SDL3/TouchFingerEvent.gen.cs
@@ -3,7 +3,6 @@
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace Silk.NET.SDL;
 
@@ -15,13 +14,13 @@ public partial struct TouchFingerEvent
     public uint Reserved;
 
     [NativeTypeName("Uint64")]
-    public ulong Timestamp;
+    public nuint Timestamp;
 
     [NativeTypeName("SDL_TouchID")]
-    public ulong TouchID;
+    public nuint TouchID;
 
     [NativeTypeName("SDL_FingerID")]
-    public ulong FingerID;
+    public nuint FingerID;
     public float X;
     public float Y;
     public float Dx;
