+++ sources/SDL/SDL/SDL3/UserEvent.gen.cs
@@ -3,7 +3,6 @@
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace Silk.NET.SDL;
 
@@ -16,7 +15,7 @@ public unsafe partial struct UserEvent
     public uint Reserved;
 
     [NativeTypeName("Uint64")]
-    public ulong Timestamp;
+    public nuint Timestamp;
 
     [NativeTypeName("SDL_WindowID")]
     public uint WindowID;
